+++ lib/aipp/parsers/LF/helpers/html.rb
@@ -2,8 +2,10 @@ module AIPP
   module Helpers
     module HTML
 
-      def cleanup!
-        html.css('del').each { |n| n.remove }   # remove deleted entries
+      def cleanup(node)
+        node.tap do |root|
+          root.css('del').each { |n| n.remove }   # remove deleted entries
+        end
       end
 
     end